Use modal coordinator for all devices
